requirements.txt
alabaster==0.7.12
alembic==1.4.2
appdirs==1.4.3
asn1crypto==1.3.0
astroid==2.4.1
astropy==4.0.1.post1
atomicwrites==1.3.0
attrs==19.3.0
Automat==0.8.0
autopep8==1.4.4
awscli==1.18.73
Babel==2.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.3
bitarray==1.2.1
bkcharts==0.2
bleach==3.1.0
bokeh==2.0.2
boto==2.49.0
botocore==1.16.23
Bottleneck==1.3.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.17
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
distributed==2.17.0
docutils==0.15.2
ecdsa==0.15
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
fsspec==0.6.2
future==0.18.2
gevent==1.4.0
glob2==0.7
greenlet==0.4.15
gunicorn==20.0.4
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
hypothesis==5.5.4
idna==2.9
imageio==2.8.0
imagesize==1.2.0
importlib-metadata==1.6.0
incremental==17.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
jeepney==0.4.2
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.6
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.32.1
locket==0.2.0
lxml==4.5.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
mock==4.0.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.1
numpydoc==0.9.2
olefile==0.46
openpyxl==3.0.3
packaging==20.1
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.3
psutil==5.7.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
py4j==0.10.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pydocstyle==4.0.1
pyflakes==2.2.0
Pygments==2.5.2
PyHamcrest==1.9.0
pylint==2.5.2
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
pyspark==2.4.5
pytest==5.4.2
pytest-arraydiff==0.3
pytest-astropy==0.8.0
pytest-astropy-header==0.1.2
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jose==3.1.0
python-jsonrpc-server==0.3.4
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.23.0
rope==0.16.0
rsa==3.4.2
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
snowballstemmer==2.0.0
sortedcollections==1.1.1
sortedcontainers==2.1.0
Sphinx==2.4.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.0
spyder==3.3.2
spyder-kernels==0.2.6
SQLAlchemy==1.3.17
statsmodels==0.11.1
sympy==1.5.1
tables==3.6.1
tblib==1.6.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.42.1
traitlets==4.3.3
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
ujson==2.0.3
unicodecsv==0.14.1
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
yapf==0.28.0
zict==1.0.0
zipp==2.2.0
zope.interface==4.7.1
